==> DART_processor/00_CorporationListProcessing.ipynb <==
import numpy as np
import pandas as pd
import os
#from pandas.io.data import DataReader

#### 파일 불러오기

data_dir = 'D:/Dropbox/Projects/CRAWLER/disclosure_downloader/'

def openFile(market_type='KOSDAQ'):
    global krx, kind, codes
    krx = pd.read_excel(os.path.join(data_dir,market_type + '_KRX.xls'))
    kind = pd.read_csv(os.path.join(data_dir,market_type + '_KIND.csv',encoding='euc-kr'))
    krx = krx[krx.columns[1:-4]]
    
    cs_krx = set(list(krx['종목코드']))
    cs_kind = set(list(kind['종목코드']))
    
    #row 갯수(회사 수)가 같은지 확인
    if len(krx) != len(kind):
        print('# of rows : KRX data and KIND data have different number of rows!!')
        return
    else:
        print('# of rows : Confirmed')
    #code 종류(속한 회사)가 같은지 확인
    if cs_krx != cs_kind:
        print('There are some problems in corporation codes!')
        return
    else:
        print('Corporation codes : Confirmed')
    
    krx = krx.sort(columns='기업명')
    krx = krx.reset_index(drop=True)
    kind = kind.sort(columns='회사명')
    kind = kind.reset_index(drop=True)
    
    if list(kind['회사명']) != list(krx['기업명']):
        print('There are some problems in corporation list!')
        return
    else:
        print('Corporation list : Confirmed\n')
        print('=========')
        print('Completed')
        print('=========')
        codes = ['0'*(6-len(str(c))) + str(c) for c in list(krx['종목코드'])]
    return

openFile()

KOSPI_MarketCap = 'KOSPI_MarketCap_170102'

cap = pd.read_csv(KOSPI_MarketCap+'.csv')
cap = cap.loc[:,['종목코드', '종목명', '현재가','시가총액','상장주식수']]
cap['현재가'] = [int(c.replace(',','')) for c in cap['현재가']]
cap['시가총액'] = [int(c.replace(',','')) for c in cap['시가총액']]
cap['상장주식수'] = [int(c.replace(',','')) for c in cap['상장주식수']]
cap['종목코드'] = ['_' + c for c in cap['종목코드']]
cap = cap.rename(columns={'현재가':'현재가_'+KOSPI_MarketCap.split('_')[-1]})

cap.to_csv(r'corporation_information_2016.csv', encoding='utf-8', index=False)

merged = pd.merge(krx, kind, on=['종목코드'])
merged['종목코드'] = codes
merged = merged.drop(['회사명', '업종_y', '상장주식수(주)', '자본금(원)'],1)
merged = merged.rename(columns={'업종_x':'업종'})
merged['업종'] = [n.replace(',', ' ') for n in list(merged['업종'])]
merged['주요제품'] = [str(n).replace(',', ' ') for n in list(merged['주요제품'])]

merged = pd.merge(merged, cap, on=['종목코드'])
merged['종목코드'] = ['_'+c for c in codes]

merged.to_csv(r'corporation_information.csv', encoding='utf-8', index=False)

merged.sort_index(axis=0,by='시가총액',ascending=False)[:10]
