==> DART_processor/02_LatestDisclosureExtractor.ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup
import numpy as np
import traceback
from tqdm import tqdm
import time
import json

corp_info = pd.read_csv(r'corporation_information_2016.csv')

#### JSON File Dictionary로 불러오기 & key는 list로 불러오기

with open('doc_info.json', 'r') as f:
     doc_dict = json.load(f)

corp_keys = list(doc_dict.keys())
corp_keys.sort()
bsn_tp = list(doc_dict[corp_keys[0]].keys())
bsn_tp.sort()

### 최신 공시 추출

def latestDisclosures(report_type_list):
    latest_disclosures = pd.DataFrame(columns=['종목코드', '기업명','보고서명','접수번호','접수일자'])
    index = 0
    for key in corp_keys:
        df_list = [pd.read_json(doc_dict[key][tp]) for tp in report_type_list]
        df = pd.concat(df_list, ignore_index=True)
        # 비어있는 Dataframe은 일단 무시 (일일요청횟수추가로 다 수집이 안 되어있음)
        if len(df) == 0:
            continue
        df = df.sort_values(by=['접수일자'], ascending=False)
        df = df.reset_index(drop=True)
        
        for ix, rw in df.iterrows():
            if '정정' in rw.보고서명 or '추가'  in rw.보고서명:
                continue
            else:
                break
        
        종목코드 = '_'+key.split('_')[1]
        기업명 = key.split('_')[0]
        보고서명 = rw.보고서명
        접수번호 = str(rw.접수번호)
        접수일자 = rw.접수일자
        
        latest_disclosures.loc[index] = [종목코드, 기업명, 보고서명, 접수번호, 접수일자]
        index += 1
    
    return latest_disclosures

def missingCorp(disclosure_df):
    global corp_keys
    
    기업명 = list(disclosure_df.기업명)
    종목코드 = list(disclosure_df.종목코드)
    existing_keys = [c1+c2 for c1,c2 in zip(종목코드)]
    
    missing_corp = list(set(corp_keys) - set(existing_keys))
    missing_corp.sort()
    
    return [c.split('_')[0] for c in missing_corp]

latest_disclosures = latestDisclosures(['A001'])

#### 정기공시 없는 리스트

corp_info[corp_info.종목명.isin(missingCorp(latest_disclosures))]

### Download Link Column으로 추가

latest_disclosures.head()

def getDocumentNumber(disclosure_df):
    base = 'http://dart.fss.or.kr/dsaf001/main.do?rcpNo='
    dcm_no = []
    for index, row in tqdm(disclosure_df.iterrows()):
        url = base + row.접수번호
        r = requests.post(url)
        time.sleep(np.random.randint(1,300)/100)
        soup = BeautifulSoup(r.text,'lxml')
        time.sleep(np.random.randint(1,300)/100)
        dn = str(soup.find_all(href='#download')).split('openPdfDownload')[1].split(';')[0].split(',')[1].split(')')[0].split()[0][1:-1]
        dcm_no.append(dn)
    disclosure_df['문서번호'] = dcm_no
    return

getDocumentNumber(latest_disclosures)

latest_disclosures.head()

def getDownloadLink(disclosure_df):
    base_pdf = 'http://dart.fss.or.kr/pdf/download/pdf.do?'
    base_xls = 'http://dart.fss.or.kr/pdf/download/excel.do?'

    pdf = []
    excel = []
    for index, row in tqdm(disclosure_df.iterrows()):
        rcp_no = row.접수번호
        dcm_no = row.문서번호
        
        pdf_link = base_pdf + 'rcp_no=' + rcp_no + '&dcm_no=' + dcm_no
        xls_link = base_xls + 'rcp_no=' + rcp_no + '&dcm_no=' + dcm_no
        
        pdf.append(pdf_link)
        excel.append(xls_link)
    
    disclosure_df['보고서링크'] = pdf
    disclosure_df['재무제표링크'] = excel
    return

getDownloadLink(latest_disclosures)

latest_disclosures.to_csv(r'latest_disclosures_KOSPI_170102.csv',index=False, encoding='utf8')

latest_disclosures.to_csv(r'latest_disclosures_KOSPI_170102.csv',index=False, encoding='utf8')
