==> DART_processor/03_DisclosureDownloader.ipynb <==
import urllib
import pandas as pd
from tqdm import tqdm
import numpy as np
import time

#links_df = pd.read_csv(r'latest_disclosures_KOSPI_160104.csv')
links_df = pd.read_csv(r'latest_disclosures_KOSPI_170102.csv')

links_df.head()

for index, row in tqdm(links_df.iterrows()):
    directory = './data/latestDisclosure_KOSPI/2016/'
    url = row.보고서링크
    name = 'KOSPI'+row.종목코드+'.pdf'
    urllib.request.urlretrieve(url,directory+name)
    time.sleep(np.random.randint(1,400)/200)

url = 'https://www.amazon.com/SAMSUNG-50-Inch-Crystal-AU8000-Built/product-reviews/B08Z1RN7NP/ref=cm_cr_arp_d_paging_btm_next_2?ie=UTF8&reviewerType=all_reviews&pageNumber=2'
